class18/pyproject.toml
[project]
name = "class18"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.3",
    "pwdlib[bcrypt]>=0.3.0",
    "pyjwt>=2.11.0",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
]
